add md/cel.go and move to related code to it
